A working ARM meta-compiler!!
Run as follows:

  ./muforth -v -f target/ARM/load.mu4

You'll see lotsa stuff fly by! The meta ! word prints out the address
and contents that are being stored.

-v shows the names (and vocab and here pointer) of words as they are
being defined.

This currently compiles a -headless- forth, but that's not too much
work. One realy cool thing is that the target-resident control structure
words (if then else etc) are included!

I have no idea if it runs (probably not) or even if the code is being
assembled correctly, but it's pretty awesome!