Meta-compilation now generates C code
Even with the debugging code still in, it's possible to get reasonable
output, since the debugging (and status) code all writes to stderr.

By doing something like this:

  ./muforth -f target/ARM/load bye > image.c

it's possible to generate a compilable C file!

I also checked (by "eye") the generated relocation bitmap, and it seems
totally correct! So, we're on our way...