Moved a few strings words around; scrabble no longer C
This is a cleanup - it introduces no new functionality, nor does it
change the semantics of any words.

Moved scrabble out of C because it didn't belong (;-) and because I
wanted to see how bad it would be write it in Forth, using cmove. This
is in preparation to creating meta-compiler code to build both
host-resident and target-(image)-resident heads.