Metacompiler now compiles names into the image
It seems to work well. I used "link>name" to crawl around and follow
links, and everything seemed fine.

I also "caught" an interesting bug - before it bit anyone. I compile
links using a, since they should be relocated, right? But we mark the
end of the chain with -zero- which must be compiled as a literal
(non-relocated)! So the linking code tests the link it's about to lay
down and makes it relocatable only if non-zero.

That would have been an infuriating bug!