Support for defining words added
This includes things like does> and ;code, and the words needed to make
them go.

I had to move a couple things around - like putting some of the register
defines -back- into the assembler (but at least in that case I think
they are normal and generic enought that they belong there).

None of this code is used yet, but the pieces are all there. Next time
I'll run them through their paces!