Proper defining words! and they're used in the kernel
Now the only forward-referenced "primitive" routine (that is, not a CODE
word, but the pieces of code that code fields point to - such as
docolon, dovar, doconst, etc) is -dodoes-.

Once I add the target-resident defining words - the ones I'm committing
now live in the meta-compiler - I'll have to solve the "what manner of
beast -is- dodoes, anyway?" puzzle. I'm punting for now. My head is
already hurting (literally and figuratively).