I think I just reached perfect Forth enlightenment!
The discussion is longer than I have time for, but suffice to say that
dodoes is both special and not special!

dodoes, unlike other bits of "primitive" code - such as docol,
dovariable, doconst(ant) - which are pointed to (or called) directly by
child (create'd) words, dodoes is what the "code field" of the -parent-
(the word containing "does>") points to (or calls.

Code field?!? Of the parent word? What does this mean?

Well, it turns out that when you think about it, and stage the
compilation of it in two stages, it all falls out.

<dodoes> which compiles the "code field" for the parent, is in the same
class - it's a -defining- word!! - as create, constant, and : .

does> calls <;code> (to capture the address of the parent, to put into
the children words) and then calls <dodoes>, which, having earlier
captured the address of dodoes (the actual machine code routine),
rewrites the parent "code field" to point to dodoes.

In this way it's possible to write all the primitive code routines in
defining words using ;code.

It's pretty amazing!