Maybe fixed rot and -rot in code?
And I added a comment in the ARM assembler about which versions of ldm
and stm you get by default.

The ldm's were doing the right thing (leaving the base alone and loading
registers upwards from the base (sp); but the stm(db) was first
-decrementing- the base by regs*4 before doing the moves) and so stored
the new values in empty space below (in the sense of towards lower
memory) the "top" of the stack. So the second and third elements of the
stack weren't affected; only top.

It might be correct now.