Added a ton of stuff to the ARM kernel
This kernel compiles but I'm not sure if what it generates is correct. I
wanted to kick it over the fence while I verify it.

A good chunk of the following parts are all included now:

  * printing numbers
  * tokenising input
  * creating new names in dictionary
  * target-resident defining words

They need to be fleshed out - and tested! - and we're still missing a
few things:

  * number input
  * dictionary search
  * interpreter loop

I had to stub out a few things to make it all compile. Words that have
names but no definition are marked UNDEF.