Number i/o and basic math routines are now double-length
I did some digging and pulled up an old version of startup.mu4 that did
proper Forth-style double number input and output conversion, and
scaling (*/) operators with double-length intermediate products. These
had been removed because of the difficulty of getting gcc to generate
remotely sensible code, and as I was hoping for some semblance of
portability when I converted muforth to threaded code, I had to remove
the (altogether simple and straightforward) implementations of the basic
multiply and divide, and make everything single-length.

That was a sad day for me.

So, today is a happy day. I've added this all back in - now living in
target/ARM/kernel.mu4. I've added stubs where necessary to unimplemented
code routines.