Added code to demonstrate automagic generation of C interface code
Once we decide it's correct, we can turn it into true asm macros!

"call" is the workhorse. You give it the number of input parameters (0
to 2) and the number of returned values (0 or 1) and it generates code,
assuming r3 holds the address of a function to call.

It's smart about only pushing or popping the data stack exactly the
rigiht amount of times!