ARM asm now allowed negative offsets in ldr/str
This makes loading PC-relative constants that are sitting in memory just
before the code (like the 'type and 'key fields) easy.

While in there I fixed (deprecated to zero) the use of macro: in the
assembler. The few (three) places I need to pull in an asm word I just
\a. macro: changes the context too much (searches .meta. as well) and
makes writing host-side forth code too hard - the switching of contexts
in your head is burdensome.