Added code to automagically generate C call interfaces
It's easy to use. Just say

  <takes> <leaves> call

where <takes> is the number of parameters the C function takes (0, 1, or
2), and <leaves> is the number it returns (0 or 1).

The generated code will optimise the pushes and pops of top.

It currently uses my blr macro so I can decompile and see what's going
on, but it's easy to switch back to blx.

In the process I realised just how screwed up macro: is. I had to change
the search order (.compiler. now before .assembler.) to make the code
work. I had tried it with strings and it worked immediately. It took
several hours to get this working - all because macro: is broken.

I'm tempted to make it go away, and instead have a "stream mode" for
assembling code inside a normal colon word. I've done this before - it's
just that I want a lighter-weight notation. I used <asm ... asm> before.
Not sure what to try...