Renamed some dictionary structure words
I had inadvertently created a word called "does>" which goes from the
cfa to the next cell, where the IP address of the does> part of the word
is. But this was immediately shadowed by the following definition of
the real "does>"!!

So I renamed >does and does> to >ip and ip>, since that's really what
they are.