Target colon compiler made much safer
Forward referenced primitives - ^ lit (branch) and their ilk - are now
checked as they are being used. If the variable that holds the value
(target cfa) of the primitive is 0 when compiled (by the word "p,") an
error is thrown and a nice message printed.

This way it is impossible to compile "bad" colon words - with 0 pointers
in their bodies.