Added cmove and refactored string=
They are structurally very similar, so I wanted to make that clear, and
make the code as readable as possible.

Right now both use "r> drop" to drop the for loop counter. Once we
realise that this works, we can substitute "rdrop" and see if that still
works.

When I tested all the dictionary code under muforth I used "r> drop"
exclusively and it was fine.