Relocated cells are now "decorated" by the memory dumper
In order that we don't get bit by any more relocation-related bugs, I
thought it would be cool to have the memory dumper show which cells are
relocated (marked in the bitmap) and which are not.

With this commit, cells that -are- relocated have a prefix `+' before
their cell-sized value.

It's very easy to see, with this change, that the values compiled into [
and -] are wrong - that they are -literals- rather than -address
literals-.