Metacompiler marks target's .forth. and .compiler. as relocated
So fixup in warm is no longer necessary.

We were bitten again by the bug that, in addition to -literals-, the
values of certain variables need to be marked as relocated. By default
the values of variables are compiled using , (rather than a, ) and so
are marked as non-relocated.

After loading the entire kernel, the meta-compiler fixes up .forth. and
.compiler. to point to the last words in each chain. Since .forth. and
.compiler. are simply variables, their values weren't marked as
relocated.

Now the word `parallels' - which does the fixing up - marks them as
relocated.