Removed public.h from all target in Makefile
Everything seems to work fine still. Right now the .o do -not- depend on
public.h, even though they all #include it (indirectly thru muforth.h).
I don't think this is a problem, since public.h is only for defining all
the available mu_ externs. If all the .o files depended on it, then any
change to any file would cause all of muforth to be rebuilt, which seems
wrong.

forth_chain.h and compiler_chain.h are -always- rebuilt if for any
reason public.h changes. This is how it should be. Which means that
dict.o (which depends on *_chain.h) will be rebuilt whenever public.h
changes.