Working double number input!
But it has the same bug my old code had - namely, that a "number"
consisting only of one of the punctuation characters ( , - / : _ ) will
convert the number 0, rather than triggering a "not defined" error.

I previously solved this by having a word "digits" that saves the
current character count before calling >number, so it can see how many
digits were converted. If none were converted, it doesn't call ?dot.

This works, but in general my number code is baroque and needs to be
radically simplified. I'm working on this next.