Added do loop +loop leave and ?leave to muforth!
I figured that there is no point in being intentionally idiosyncratic
or being curmudgeonly about for-next being better or simpler than
do-loop. I'm not sure I believe that anyway.

I decided to put it into muforth, and now that I know how, it'll be easy
to put it into target forths.

Amazingly, the code worked the first time. ;-)