Moved interpret into Forth
This is a big change. It introduces what is essentially another level of
bootstrapping. The reason for doing this is that the C interpret loop
won't notice changes in depth of RP nesting; thus we can't use nifty
return stack tricks to control the system.

I'm trying to get rid of setjmp & longjmp, and catch-throw, and replace
it with something more elegant and flexible - namely, Klaus Schleisiek's
elegant nesting of vari-purposed "frames" on the R stack.

In order for this to work, changes in the depth of the R stack have to
be reflection in the execution on the C side. The way that execute_xtk()
is written - looping until RP < RPsaved - will do this - BUT ONLY IF
there are no indefinite loops in C!

And that's exactly what interpret() was.

So we have to replace it.

In earlier revs of muforth I replaced the C token consumers for
interpreting and compiling with ones written in Forth. But my main
reason for doing this was to add number parsing. I still need to add
number parsing during the execution of startup, but since I'm going to
be replacing those token consumers and rewriting interpret and evaluate
-anyway-, and I don't want to do it -twice-, now I define the C token
consumers with slots for number and number-comma, and patch them in once
muforth can understand numbers.

At the end of startup I throw away the C versions entirely, and replace
them. And I have to replace : ; [ -] and a few other words as well...

Since catch/throw is gone, I made die() a bit friendlier. Now if an
error occurs during the reading of startup, a token, line number, and
message are printed.

Right now error handling is broken, since die() still gets called, but
it is no longer getting tokens from interpret...

Work to do still!