Cleaned up ;code in metacompiler - code & comments
Now ;code is a metac construct. It compiles (assembles) machine code.

Also defined ;calls - it's for defining the target-resident definers
that will -share- the code laid down by the metac definers... hence the
name ;calls - a word defined by the target-resident "variable" "calls"
the already-compiled code for dovar.

It's hard to explain, but easy to understand. Really. ;-)