TOP is now simply SP[0]
I decided that having TOP be a separate variable - rather than truly the
top of the stack (SP[0]) was pointless complexity. There is even the
argument that the code generated by the compiler suffered...

With TOP the same as SP[0] the code is a bit simpler. Shockingly, very
few changes were necessary, and many of them made the code even more
readable.