The beginnings of editline support - in Forth!
I was inspired by Dave's CFE demo... I've wanted command history in
muFORTH for ever!

I looked at how Mitch did it in CFE and took inspiration from that. At
first I thought of doing it in C (might still) but figured it couldn't
be -that- hard in Forth, so I started. It's very incomplete and rather
buggy, but it's a beginning.