Use uname -m rather than uname -p to figure out cpu
./configure.sh was recently changed to try to figure out if muforth is
being built on a 64-bit platform, and, if so, to add flags for gcc and
ld to specify building it as a 32-bit executable.

For non-Darwin systems I was testing the result of "uname -p" - which
queries the "processor" - but that yields odd results like "Pentium III
(Coppermine)" rather than what I want, which is "i686" or "x86_64" or
something like that.

uname -m (which queries the "machine" type, whatever *that* means, and
the GNU info page for uname is no less mysterious than the man page)
seems to give better results.

I need to try this on several Linuxes and on Net- and FreeBSD to be
sure. I'm hoping for now that using -m is better than -p.