Minor fixes to test-double-math
Changed a few "long comments" to match the newer, longer long comment.
Added a bit of debug code for bracketing number input. Also changed a
definition to align with new "traditional" does> r.t. old muforth-style
constant/does>.

Note that this code is now out of sync with the target/ARM/startup
number input code, which changed to counting the digits converted in a
variable rather than on the R stack. It _might_ make sense to bring that
change to this code - and then to muforth's startup code!