On the way to adding trap frames...
This commit builds but doesn't run. In the middle of the conversion to
re-defining the interpreter and compiler token consumers _and_ the main
interpret loop, all at once. But we need trap frames, and we need to
revert a change that re-used the C-based token consumers. Those two are
next.