Better factoring of platform-specific io
The first time around I was a bit too aggressive. I didn't need to throw
away >stderr and >stdout. Now, instead of having "type" be a core io
primitive, we keep "write" (which takes an fd) and let the platform
decide to use or ignore the fd passed to write. This way the high-level
code can decide where it wants to write text, and the underlying
platform can decide if it's possible to make distinctions, or if
everything simply goes to the console.