A better way of doing platform-specific compilation
Now, instead of lots of WITH_FOO #defines, the makefile defines certain
object files to be compiled and linked for each target platform. (There
is one WITH_ sitting around still, but it is going to go away.)

It's also much easier to specify the platform you want, once, and it's
out of the way (not under version control). The makefile tells you how.