Hack! public.h now a dependency of "all" target
It's not clear to me where to put public.h. If all the objs depend on it
(remember: every .c includes muforth.h, and muforth.h includes public.h)
then any time _any_ C file is changed, a new .ph is generated, which
generates a new public.h... which would then force the recompilation of
_all_ of objs. This is just as good as not having a makefile at all!

But removing public.h as a dependency of all the objs, and not putting
it somewhere else, means that after a make clean it doesn't _ever_ get
remade. This is no good as well.

As a hackish solution I put it as the _first_ dependency of "all"
(before muforth). Now, if it doesn't exist, it is made, then muforth is
made (based on its dependencies). Any time a .c changes, it triggers a
.ph which makes a new public.h, so it should stay up to date.

I think this is a better solution.