main() and command line code moved into a separate file (nix-main.c)
Now muforth is callable more like a library. While it still uses global
variables (they are not encapsulated into a struct, like in Lua) at
least it's easier to embed muforth in something else, because it can be
compiled without a main() routine and with no reference to a command
line.