Dictionary is now hashed 32-ways!
Instead of having each vocab chain be an entirely separate linked list,
now there are 32 lists, and the words of all chains are intermingled -
but they are also "hashed", so it's still possible to distinguish
between .forth. and .compiler. words, eg.

It's very simple. When a new word is defined, it is put on chain h,
where h is an index from 0 to 31, calculated as follows:

  h = ((first letter of name) + (length of name) + (vocab index)) mod 32

.forth. has index 0; .compiler. index 1, and as they get defined, other
chains get increasing indices, up to 31. (With 32 heads there can only be 32
independent chains.)

Likewise, when searching for a word, we hash it and look on chain h. If
it's not found there, it doesn't exist _in that vocab chain_. There
could be a word with the same name on list h+1 but that belongs to a
different chain!

So everything works great. With big vocab chains loading should be quite
fast.