Back to original interpret mode implementation!
Well, not exactly.

The major purpose in all of this reshuffling is to get rid of setjmp,
longjmp, catch, and throw, and to do error handling in a more
Forth-like and elegant way. The major impediment was one C routine:
mu_interpret(). The problem? It contained an indefinite loop, thus
preventing us from controlling its execution by simple Forth-like means,
such as resetting IP or popping the R stack.

My original solution to the problem was much too complicated. Instead,
now I've gone back almost to my original model - state, [, and -] are
defined in C and the "invisible" routines that interpret and compile
tokens are patched (the way they always were) once number input has been
defined.

The difference? Now, at the end of startup.mu4, we also redefine
interpret. But it has the exact semantics of the C version! But because
it's in Forth we have more control over its execution.

I almost have a (kludgy) version of nested loading working. Once that's
working, I'm going to patch in the real R stack code!