Real R-stack-based error recovery works!
Now, when there is an error, right then, at that moment, information
about where (file and line number, if not interactive) and what (the
error message) is printed, and only *then* is the stack unwound and
variables reset. This means that there no need to do kludgy "capturing"
like I did with catch/throw. This is much simpler, and much nicer, and
works great!