file[# and #]file are replaced by loading"
The refactoring is much nicer, much more robust, and takes advantage of
our R-stack nesting and recovery code.

When a file wants to be loaded verbosely it calls loading" with a string
describing itself. loading" prints (( and the text, then saves here on R
and recursively calls interpret, setting on-exit to call the routine
that calculates dict space and prints the closing )).

It's really elegant and simple.

Unfortunately I couldn't get stack depth checking to work as neatly.