S08 to S08 BDM bit routines use fewer resources now
Only one pin and one timer channel! It's tricky getting the TPM to share
with the regular pin driver.

I also made a small change in the 8M tx1 bit routine - its start bit is
now 5 cycles, rather than 4, and better matches the rx1 routine.
Probably not a big deal, but I wanted them to be the same.