Minor usability tweaks in chat/bdm code
Two changes:

(1) no longer wait for a byte when switching to bdm mode (this is still
broken - switching needs a better solution).

(2) don't switch the dump c@ word until we're actually chatting with a
target. So it's possible to look at the bytes in the image, locally on
the host, before connecting (or without connecting at all).