Tiny changes to 908chat code
Using clr rather than mov #00; also put in a comment to prevent me from
changing the first line of the S08 reset code, which disables the
watchdog, from three instructions to a single bclr. It _has_ to be three
instructions: the register being changed is _not_ in the zero page.