Major rework of memory dump code
This has gone thru several iterations over the years. It's still far
from perfect (in fact, I'm checking it in now because it's making me
crazy and I don't want to think about it any more - for now, anyway!)
but I think it's simpler.

Part of the complexity of "getting it right" is that I want to share the
core code between dumping host memory, and
dumping/disassembling/decompiling target memory. Doing that seems to
involve lots of deferred words, and it simply starts to get ugly. This
code at least works, and seems reliable. When I'm in a better mood about
it I'll revisit it.