Hacked target/HC08 to work with new memory dump code
Disasm works - sort of - and memory dumping (with "interlaced" image and
target lines, for easy comparison) works with "new" lib/see.mu4.

All of this is a bit of a crock, and still needs to be simplified and
reworked, but I've been tinkering with it for several days and haven't
found the sweet spot, and I'm sick of looking at it! So, here it is, in
all its imperfection.