Slightly improved multi-platform support
Renamed nix-main.c -> main.c
Changed default platform to "nixy"; this is the Unixy default. Added a "nixed"
platform -- this is nixy without the trimmings (time, tty, select).

Added a muforth-bare target to see what symbols are undefined when we don't
link to the default libs. This is useful to see what our platform dependencies
really are.