Dictionary allocated with calloc, not mmap
This should be more portable. I'm still using mmap to read files. The trouble
with mmap wasn't with the file reading but with MAP_ANON being something that
needed special non-ansi settings to make available. It was also different
between Linux and OSX in terms of what was deprecated or encouraged.

With a big enough chunk of memory it's likely (Linux libc even specifies this)
that malloc/calloc will simply call mmap behind the scenes to do what I was
having trouble doing portably.

I think this is an improvement. ;-)