There is now only one comma (compile, is gone)
Since there is only one dictionary space now (it's been this way for a while)
having  ,  and  compile,  seemed redundant and silly.

I removed  compile,  and in the process slightly streamlined the set of basic
dictionary operations. Now  ,  allot  here  are all defined in C.  allot
always aligns, so there is *no* way to cause the dict to get misaligned. One
way we prevent this is that  h  is no longer exported outside of C.

Even so, I've added an assert so that every time a new name is defined, the
dict alignment is checked.