Using "smart" @ and ! to simplify variable use
The problem: on 32-bit machines, the C variables are all 32-bits -
"address-sized" - and so we need to use a@ and a! to access them. It's
annoying to have to get that right in the Forth code. It's esp annoying, since
in order to use  preserve  with addr-size vars, we have to create a _copy_ of
it - called  apreserve  - that uses a@ and a! internally. This is a huge hack,
but it was necessary to get the 32/64 hybrid version working.

This fixes that problem. By making  @  and  !  smart about the address that
they are passed ("is it inside the Forth heap?") the programmer (and preserve)
can simply use @ and ! and the right thing will happen.

Note that access to colon word bodies still needs a@ and a!, so they don't go
away completely. But those usages should be limited to very deep technical
changes (like changing how control structures, or create/does, work) and so
shouldn't pose a problem.