Vocab chain "heads" are now defined dynamically
This method improves over previous methods, but it's still not perfect. For
one, "hiding" names is a bit kludgey. Currently I'm just rewriting the first
character with a DEL. But better would be to take the high bit of the length
as a smudge bit, a la fig-FORTH. (As I've said earlier, muFORTH asymptotically
approaches fig as time goes by...)

These heads are defined in such a way that it's possible to "chain" chains
together, so that a search in .foo. continues in .forth., if .foo. has been
chained to .forth. . Because of this chaining, the names themselves shouldn't
show up ... or should they? In some ways it can be nice to see where the vocab
structure branches join. So there are other options, and this may change in
the future.