Flags in names are gone
I was wrong. It's too much.

So I've decided to "get my fig on" and use the high bit of the length byte as
a "hidden" flag. This is much nicer, and simpler, and works great. I'm also
now limiting the length of names (to 127) whereas before there was no limit...
which is just plain wrong.