A pure 64-bit version! Even on 32-bit machines!
Ok, so this is part of a grand experiment - or a series of them - not
yet finished.

In this checkin, everything that can be is 64-bit:
  * both stacks
  * code, ip, and link pointers
  * docolon word code bodies

What this means is that on 32-bit machines, compiling the same Forth code
takes almost twice the space. But you get a uniform "sense" of being on
a 64-bit machine. The main perq is the 64-bit math support.

One of the tricky bits was hiding, on 32-bit machines, the 32-bitness of
the underlying C code. Several C variables, for instance, are exported
to Forth, which wants to use @ and ! to read and write them - but it's
going to read or write 64 bits, but the actual C value is only 32. Havoc
ensues. So I "wrapped" each of these values in a struct to make them 64
bits long. C uses only the half it cares about, while Forth can happily
fetch and store.

Now, there is a lot still to be done, and I may still make a hybrid
32/64 version available - which, on 32-bit machines, will have 32-bit
pointers in the dictionary, and 32-bit docolon bodies, but all Forth
values and both stacks will be 64-bit. This was the first version of
64-bit muforth on 32-bit machines, but aspects of the hybridness of it
bugged me. I think I may have a better solution.

For now though, 64-bit purity is the word of the day.