First pass at new approach to 64-bits
This is rev four or so...

I realised today that my approach - esp to making hybrid 32/64 versions
- but even in making a "pure" 64-bit version that will run on 32-bit
hardware - was kind of "inside-out" from how it should be. Or, if not
_should_, then at least _could_.

Instead of going thru elaborate machinations to "cell-wrap" C variables,
and use macros to parametrise both C var access and the actual
"execution engine" bits of the code (so that it can work "transparently"
on both 32-bit and 64-bit hardware), another approach is to let the
hardware be itself. On 32-bit machines, let addresses be 32-bit, and use
them natively in the execute engine. Likewise for 64-bit.

Also, since the whole reason for doing this grand experiment was to be
able to compute easily with 64-bit numbers, keep the width of values on
the stack at 64 bits. Addresses are their natural width. But what
happens if we _mandate_ (even on 64-bit platforms!) that the Forth cell
size is 32 bits?

That's what I've done in this checkin. I'm suddenly realising that maybe
it's not such a great idea, because now we have to be really careful
about distinguishing addresses from values. I was even thinking of a
clever tagging system, so that stack values would be tagged, and the tag
could even such metadata as "this is a pointer to a pointer to data"...

But isn't it easier if cells and addresses are the same size - the
natural word width of the machine - and the stack is always 64 bit?
Doesn't that give us _exactly_ what we want?

I have to think about this some more. ;-)