Cells and addresses are the native word size; stacks are 64-bit
This combination seems pretty sweet. The code is just as simple as
muforth traditionally was. The only slightly odd thing is that access to
the stacks via _address_ (ie calculate of of sp@ and then fetch or
store) is awkward. I created the words s@ and s! for this: they will
_always_ read or write 64 bits, unlike @ and !, which operate on the
native word size.

Next step: to try the code on a 64-bit machine!