All objects depend on public.h again
But, thanks to a suggestion by Michael Pruemm, this doesn't mean that
everything is rebuilt all the time. Instead, the .ph files are cat'ed
into public.h.new and only if this _differs_ from the existing public.h
is it copied over. Thus public.h's timestamp changes - forcing a
recompile of _all_ .o files - only when it's contents have actually
changed.

If Make used a better dependency algorithm - like SHA1 hashes - to see
when files have changed, silly hacks like this would not be necessary.