Moved current (new) new and : out of dict.c and into startup.mu4
Currently the two primitives for creating new names in the dictionary
are:

  (hidden-name)  link a u -- 'suffix
  (linked-name)  a u chain --

(hidden-name) creates a hidden name (!!) in the dictionary, using the
passed link as its link field and the token (a u) as the name; it
returns the address of the suffix cell of the new name. This is used -
exclusively so far - for creating new vocab chains.

(linked-name) is the "norrmal" way of creating new words. It takes a
token (a u) and a chain, creates a non-hidden name, sets its link to
point to chain, and sets chain's link to point to the new word.