(name) and (linked-name) are the two creation primitives now
A very slightly different factoring. (linked-name) is still what is used
for normal words.

(name) is the fundamental word that creates new dict entries. It can
create hidden or non-hidden names, but while it fills in the new word's
link field (with a passed-in value), it doesn't link the name into any
lists.

Again, currently the only use is to make new vocab chains. But at least
now the more general mechanism is exported to forth-land.