Unbroke macro compilation
Way back in commit 0ccab22a I changed the order that dict chains get
searched when compiling macros, and broke some macros in the BDM code.
I've now reverted that part of the change. The assembler's
begin/if/while etc now _trump_ the colon compiler's - this is the way it
_used_ to be.