Converting chat code _back_ to 115200 bps!
As is (too) often the case, I'm making work for myself. I was concerned
that 9.216M wasn't "fast enough" (whatever that means) and in trying to
eke out a bit more speed (up to 9.8304M!) I decided to slow the baud
rate for the serial comms.

I hadn't yet done the conversion of the S08 BDM, which would have meant
minor rewriting of the bit bang routines. (It's a good thing I didn't
get to it.)

This is all very silly. These parts, at these speeds, are only going to
be used for BDM. For any other use, we can set the serial clock rate at
will.

To paraphrase my last commit message:

So I'm switching to 9.216M. It's fast enough. This all goes away once
we've got USB working anyway.