Cleaned up S08 flash code and treatment of stack frames
Oddly, they are related. I started this because I wanted to combine the
chat and bdm "modes" in a more seamless and reliable way. One issue I
had was that the structure of the current chat code requires that
command processing _not_ be in a subroutine, so the stack frame is the
same as the SWI frame. But this only matters if you want to _set_ SP...
So I looked for where I set SP... and went down a rabbit hole!

I haven't removed the SetSP command from chat, but now I can. I've
restructured the way stack frames work a bit too - but it's still
subobtimal. Esp now that we are "scripting" the process of writing to
flash, when in chat mode, pushing and pulling the redundant copy of the
SWI frame - for every byte to flash is silly. So that could be fixed.
Probably only BDM mode needs its own host-side copy of the registers.

Speaking of BDM mode... it's now possible to write to the flash on a
remote chip (accessed via bdm) without downloading any code into RAM!
All the operations are "scripted" via the BDM interface.

The flash code now reflects new factoring. There are "setup" and "run"
words for chat and bdm - two deferred words to run them. (And you still
have to select which you want now - it's not automatic. You have to say
"flash-via-chat" or "flash-via-bdm".) But overall the flash code is
better, and the RAM-resident bit for flashing via chat is much shorter
and simpler.

I also learned something deep esoterica about the flash machinery and
the charge pump. ;-) I was trying to use the burst program mode - which
pipelines the entry of flash commands (as long as they are all Burst
Program) and doesn't turn off the charge pump between bytes. It takes
4us instead of 9us to program a byte. (Not that this matters to me - the
comm overheads of my setup _far_ outweigh those times. I just wanted to
do it the right, elegant way. ;-)

So here's the trouble: you jump to your RAM routine to write the byte
into flash. It stores it into the array, sends the right commands to the
right registers (and the flash is written correctly), and exits when
it's okay to send another (burst) program command. But the flash array
isn't accessible, and the charge pump is still going! But your routine
does an RTS back to flash... and the chip resets with an illegal
instruction.

Changing the code to use the non-burst program command instead solved
the problem.

Phew!