Host's du now renamed du-host; du becomes target du
Having one in .forth. and one in .meta. wasn't working for me. I'm not
reliably ending up in __meta mode, and when I type "0 du" in __forth
mode, I crash. This is annoying. Now, it doesn't matter - in either
mode, once all the code is loaded, du is the target du.