Changed to chat2 protocol
And renamed c.Go to c.Run, and c.Run to c.RunWait. More memorable names.

The changes here are twofold:

 * chat no longer defines an Idle command; anything not defined is by
   definition an Idle; Bye is still 0.

 * when the target resets, and when it hits an swi and returns to the
   chat loop, it no longer sends a byte.

This second one is a bit subtle, but makes sense. There is no longer a
distinction between Run (just send the target off running) and RunWait
(start it, then wait for completion) _at the protocol level_. It has to
be done using the other commands in the protocol. But this is fine (and
it's what BDM does naturally), since all you have to is request the
value of the stack pointer, and you won't get it until the target
returns to its command loop.

As a help to upgraders: to upgrade a 908 part from chat1 to chat2, do
the following:

  ./muforth -d chat-in-ram -d chat1-host -f target/HC08/load-908boot1.mu4
     chat
     hi
     chat-go
     ^C
  ./muforth -f target/HC08/load-908boot2.mu4
     chat
     flash-image
     flash-vectors

For an S08 do this:

  ./muforth -d chat1-host -f target/HC08/load-S08chat0.mu4
     chat
     hi
     chat-go
     ^C
  ./muforth -f target/HC08/load-S08chat2.mu4
     chat
     flash-via-chat
     flash-image
     flash-trims
     flash-vectors

It's also possible to do using an "intermediate" flash-resident version
(which, like the ram version, uses an emulated-swi) - but this flash
version isn't in the same sector as the trims and vectors. To do this
run:

  ./muforth -d chat1-host -f target/HC08/load-S08chat1.mu4
     chat
     hi
     flash-via-chat
     flash-image
     chat-go
     ^C
  ./muforth -f target/HC08/load-S08chat2.mu4
     chat
     hi
     flash-via-chat
     flash-image
     flash-trims
     flash-vectors

NOTE: this means that when reburning the chat byteloader on an S08
via chat, we cannot use SWI instructions! So be careful...