Fixed incomplete QE8 definition; also fixed .ifdef for .meta. words
I had completely left off the high register (1800+) definitions for
QE4/8.

And in making the chat code generic I broke it - by doing

  .ifdef ICSTRM

to see if this was a device with an ICS, rather than MCG, clock. The
trouble is that .ifdef only looks in .forth.! And all these words are in
.meta.

So I refactored .ifdef (it was already factored to use .def) and now
it's possible to search any chain and do conditional compilation
depending on what you find, like this:

.meta. .contains ICSTRM .if
   -- code for ICS
.else
   -- code for MCG
.then