Another slightly different "chat2" byteloader protocol ;-)
Instead of calling it chat3, I upgraded all my boards and then made all
evidence of the earlier chat2 go away.

It's still possible to upgrade from chat1. The procedure is much like
that explained in the previous "chat2" commit.

But that's not what's interesting about this commit. It contains some
huge improvements to the way that flashing works, both for 908 and S08
chips.

In both cases, the code is more correct about blank checking and erasing
pages. Before, it would check per "line" to be programmed if it was all
blank, and if so, go ahead and program. But this doesn't work very well
if there are two blank lines, followed by a non-blank one, in the same
erase page. Then, when it gets to the non-blank line, it has to erase
the whole page - erasing the two lines it just programmed. This is just
wrong. And this approach also makes handling the vectors, trims, and
security bytes more difficult.

Now, on both 908 and S08, a page is blank checked when we try to program
its first byte. And if that page (on S08) is the last page - which
contains the vectors, trims, and security bytes - then some of that info
is saved to the (host-side) image, the whole page is erased, and then
the vectors &c are programmed again - __before__ the line that triggered
the blank check. It all works very well.

One other thing: it's now possible to program the 908 flash in "forced
monitor mode". This makes it possible to rescue (some) b0rked chips. I
was able to mass erase (and reprogram the trim byte) on a chip that
seemed lost to us.

One editorial comment about all of this: the 908's are _much_ harder to
deal with than S08's. I'm glad I'm not focussing on using them!